Moves the existing renderers to the Aztec libraries.
