+++ Aztec/Classes/Renderers/HTMLAttachmentRenderer.swift
@@ -1,15 +1,14 @@
 import Foundation
 import UIKit
-import Aztec
 
 
-// MARK: - CommentAttachmentRenderer: Renders HTML Comments!
+// MARK: - HTMLAttachmentRenderer: Renders Unknown HTML
 //
-final class CommentAttachmentRenderer {
+final public class HTMLAttachmentRenderer {
 
     /// Comment Attachment Text
     ///
-    let defaultText = NSLocalizedString("[COMMENT]", comment: "Comment Attachment Label")
+    let defaultText = NSLocalizedString("HTML", comment: "HTML Attachment Label")
 
     /// Text Color
     ///
@@ -22,7 +21,7 @@ final class CommentAttachmentRenderer {
 
     /// Default Initializer
     ///
-    init(font: UIFont) {
+    public init(font: UIFont) {
         self.textFont = font
     }
 }
@@ -30,16 +29,16 @@ final class CommentAttachmentRenderer {
 
 // MARK: - TextViewCommentsDelegate Methods
 //
-extension CommentAttachmentRenderer: TextViewAttachmentImageProvider {
+extension HTMLAttachmentRenderer: TextViewAttachmentImageProvider {
 
-    func textView(_ textView: TextView, shouldRender attachment: NSTextAttachment) -> Bool {
-        return attachment is CommentAttachment
+    public func textView(_ textView: TextView, shouldRender attachment: NSTextAttachment) -> Bool {
+        return attachment is HTMLAttachment
     }
 
-    func textView(_ textView: TextView, imageFor attachment: NSTextAttachment, with size: CGSize) -> UIImage? {
+    public func textView(_ textView: TextView, imageFor attachment: NSTextAttachment, with size: CGSize) -> UIImage? {
         UIGraphicsBeginImageContextWithOptions(size, false, 0)
 
-        let message = messageAttributedString()
+        let message = messageAttributedString(with: attachment)
         let targetRect = boundingRect(for: message, size: size)
 
         message.draw(in: targetRect)
@@ -50,8 +49,8 @@ extension CommentAttachmentRenderer: TextViewAttachmentImageProvider {
         return result
     }
 
-    func textView(_ textView: TextView, boundsFor attachment: NSTextAttachment, with lineFragment: CGRect) -> CGRect {
-        let message = messageAttributedString()
+    public func textView(_ textView: TextView, boundsFor attachment: NSTextAttachment, with lineFragment: CGRect) -> CGRect {
+        let message = messageAttributedString(with: attachment)
 
         let size = CGSize(width: lineFragment.size.width, height: lineFragment.size.height)
         var rect = boundingRect(for: message, size: size)
@@ -64,7 +63,7 @@ extension CommentAttachmentRenderer: TextViewAttachmentImageProvider {
 
 // MARK: - Private Methods
 //
-private extension CommentAttachmentRenderer {
+private extension HTMLAttachmentRenderer {
 
     func boundingRect(for message: NSAttributedString, size: CGSize) -> CGRect {
         let targetBounds = message.boundingRect(with: size, options: [.usesLineFragmentOrigin, .usesFontLeading], context: nil)
@@ -73,12 +72,15 @@ private extension CommentAttachmentRenderer {
         return CGRect(origin: targetPosition, size: targetBounds.size)
     }
 
-    func messageAttributedString() -> NSAttributedString {
+    func messageAttributedString(with attachment: NSTextAttachment) -> NSAttributedString {
         let attributes: [NSAttributedStringKey: Any] = [
             .foregroundColor: textColor,
             .font: textFont
         ]
 
-        return NSAttributedString(string: defaultText, attributes: attributes)
+        let htmlAttachment = attachment as? HTMLAttachment
+        let displayText = htmlAttachment?.rootTagName.uppercased() ?? defaultText
+
+        return NSAttributedString(string: "[\(displayText)]", attributes: attributes)
     }
 }
